Various enchancements:
- Fixed `get_instruction_addresses` (not all CODE segments were detected)
- Fixed remaining Qt float/int related issues
- Updated arm32 qemu-tracer's README
- Qemu arm32 tracer - fixed memory reading

+++ plugins/tenet/ui/hex_view.py
@@ -32,7 +32,7 @@ class HexView(QtWidgets.QAbstractScrollArea):
         fm = QtGui.QFontMetricsF(font)
         self._char_width = int(fm.width("9"))
         self._char_height = int(fm.tightBoundingRect("9").height() * 1.75)
-        self._char_descent = self._char_height - fm.descent() * 0.75
+        self._char_descent = int(self._char_height - fm.descent() * 0.75)
 
         self._click_timer = QtCore.QTimer(self)
         self._click_timer.setSingleShot(True)
